requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0
ultralytics>=8.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.65.0